refactor(routes): extract leaf/domain routers from api.ts (SHR-177)
Splits api.ts into per-domain Express.Router() modules under server/routes/:
- _shared.ts — loadTaskOrFail, sendDomainError, lookupExistingReviewId,
  deepMerge, fetchTaskBundle, derivedStatus, augmentDashboardSettings
- misc.ts (8) — health, harnesses, browse, branches, preflight/*, default-branch, recent-repos
- learnings.ts (2) — /api/repos/:repoPath/learnings, /api/learnings/:id
- skills.ts (5) — /api/skills*
- teams.ts (3) — /api/teams*
- setup.ts (3) — /api/setup/*
- settings.ts (5) — /api/settings, /api/repo-config*
- hooks-registry.ts (4) — /api/hooks/templates, /api/hooks/install, /api/hooks/registry*
- chats.ts (6) — /api/chats*, /api/agents/:id/task
- agent-defs.ts (5) — /api/agents* (definition CRUD)
- orchestrator.ts (6) — /api/orchestrator/conversations*
- integrations.ts (7) — /api/integrations*

Each router uses full /api/... paths (bare app.use()) so paths are byte-identical.
Route inventory diff is empty (101 routes preserved). 47 routes remain in api.ts
for later passes (tasks, reviews, comments, diffs, worktrees).
