+++ server/api.ts
@@ -1,16 +1,10 @@
 import type { Express, Request, Response } from 'express';
 import { nanoid } from 'nanoid';
 import fs from 'fs';
-import path from 'path';
-import os from 'os';
-import { fileURLToPath } from 'url';
-import { octomuxRoot } from './octomux-root.js';
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-import { getDataDir, pingDb } from './db.js';
-import { childLogger } from './logger.js';
-import { getNeedsYou, getActivity } from './inbox.js';
 import { execFile as execFileCb } from 'child_process';
 import { promisify } from 'util';
+import { childLogger } from './logger.js';
+import { getNeedsYou, getActivity } from './inbox.js';
 import {
   startTask,
   closeTask,
@@ -22,7 +16,6 @@ import {
   createUserTerminal,
   createShellTerminal,
   closeShellTerminal,
-  hopAgent,
 } from './task-runner.js';
 import * as diffMod from './diff.js';
 import { taskWorkingDir } from './task-paths.js';
@@ -46,7 +39,6 @@ import {
   seedInlineComment,
 } from './repositories/inline-comments.js';
 import { computeOutdated } from './inline-comments-outdated.js';
-import { createChat, listChats, getChat, closeChat, deleteChat } from './chats.js';
 import { sendMessageToAgent } from './tmux-input.js';
 import { listReviewsInbox, getReviewDetail } from './reviews-inbox.js';
 import {
@@ -56,59 +48,15 @@ import {
   seedReviewRun,
 } from './repositories/review-runs.js';
 import { listPublishedReviews } from './repositories/published-reviews.js';
-import {
-  listLearningsForRepo,
-  deleteLearning,
-  addLearning,
-} from './repositories/review-learnings.js';
+import { addLearning } from './repositories/review-learnings.js';
 import { updateCommentFields } from './repositories/inline-comments.js';
-import {
-  createConversation,
-  getConversation,
-  listConversations,
-  listMessages as listOrchestratorMessages,
-  setGlobalMonitor,
-  clearGlobalMonitor,
-  getGlobalMonitorConversation,
-  getConversationUsage,
-} from './orchestrator/store.js';
-import { startConversation } from './orchestrator/runner.js';
 import { mountArtifactEndpoint } from './orchestrator/artifact-endpoint.js';
 
-import { listSkills, getSkill, createSkill, updateSkill, deleteSkill } from './skills.js';
-import {
-  listIntegrations,
-  getIntegration,
-  createIntegration,
-  updateIntegration,
-  deleteIntegration,
-} from './integrations/store.js';
-import { listProviders, getProvider } from './integrations/registry.js';
-import { maskIntegration, mergeMaskedConfig } from './integrations/mask.js';
-// Side-effect: ensure all providers are registered when the API is loaded.
-import './integrations/index.js';
-import {
-  listAgents,
-  getAgent,
-  saveAgent,
-  resetAgent,
-  createAgent,
-  deleteAgent,
-  isBuiltInAgent,
-  syncAgents,
-} from './agents.js';
 import { validateAgentName } from './harnesses/types.js';
-import { listHarnesses, getHarness } from './harnesses/index.js';
-import { getSettings, updateSettings, type OctomuxSettings } from './settings.js';
-import {
-  getOrCreateRepoConfig,
-  updateRepoConfig,
-  listRepoConfigs,
-} from './repositories/repo-config.js';
+import { getHarness } from './harnesses/index.js';
 import { hookRoutes } from './hooks.js';
 import { broadcast } from './events.js';
 import { generateTitleAndDescription } from './title-gen.js';
-import { invalidateHookEnabledCache } from './hook-dispatcher.js';
 import { ensureHookToken } from './hook-token.js';
 import type {
   CreateTaskRequest,
@@ -116,10 +64,8 @@ import type {
   AddAgentRequest,
   Task,
   Agent,
-  DerivedTaskStatus,
   UserTerminal,
   RunMode,
-  CreateChatRequest,
   MoveTaskRequest,
   SummaryRequest,
   NoteRequest,
@@ -148,10 +94,7 @@ import {
   getTaskExternalRef,
   upsertTaskExternalRef,
   deleteTaskExternalRef,
-  findReviewTaskByPrNumber,
-  findReviewTaskBySource,
   findExistingReviewTask,
-  countRunningTasks,
   getWorktree,
   listWorktrees as listWorktreesRepo,
   listTasksByWorktree,
@@ -160,7 +103,6 @@ import {
   unlinkWorktreeFromAllTasks,
   insertWorktree as insertWorktreeRepo,
   updateWorktreeFields,
-  getAgent as getAgentRepo,
   getAgentByIdAndTask,
   listAllAgents,
   listAgentsByTasks,
@@ -169,13 +111,9 @@ import {
   listUserTerminals,
   listUserTerminalsByTasks,
   getUserTerminalByIdAndTask,
-  listActiveRepoPaths,
   listTrackedRepoPaths,
-  listRecentRepoPaths,
   insertWorktreeIfAbsent,
   insertTaskIfAbsent,
-  getHookEnabled as getHookEnabledRepo,
-  upsertHookSetting,
   inTransaction,
 } from './repositories/index.js';
 
@@ -188,307 +126,45 @@ import {
 } from './services/review-service.js';
 import { ServiceError } from './services/errors.js';
 
+import { router as miscRouter } from './routes/misc.js';
+import { router as learningsRouter } from './routes/learnings.js';
+import { router as skillsRouter } from './routes/skills.js';
+import { router as teamsRouter } from './routes/teams.js';
+import { router as setupRouter } from './routes/setup.js';
+import { router as settingsRouter } from './routes/settings.js';
+import { router as hooksRegistryRouter } from './routes/hooks-registry.js';
+import { router as chatsRouter } from './routes/chats.js';
+import { router as agentDefsRouter } from './routes/agent-defs.js';
+import { router as orchestratorRouter } from './routes/orchestrator.js';
+import { router as integrationsRouter } from './routes/integrations.js';
+
+import {
+  loadTaskOrFail,
+  lookupExistingReviewId,
+  deepMerge,
+  fetchTaskBundle,
+  derivedStatus,
+} from './routes/_shared.js';
+
 const execFile = promisify(execFileCb);
 const apiLogger = childLogger('api');
-const healthLogger = childLogger('health');
-
-/** Load a task by :id param; respond 404 and return null if missing. */
-function loadTaskOrFail(req: Request, res: Response): Task | null {
-  const task = getTaskRepo(req.params.id as string);
-  if (!task) {
-    res.status(404).json({ error: 'Task not found' });
-    return null;
-  }
-  return task;
-}
-
-/** Map a thrown domain error to an HTTP response. */
-function sendDomainError(res: Response, err: unknown): void {
-  const e = err as NodeJS.ErrnoException;
-  const msg = e.message || 'Unknown error';
-  if (e.code === 'ENOENT' || msg.includes('not found') || msg.includes('does not exist')) {
-    res.status(404).json({ error: msg });
-  } else if (msg.includes('already exists')) {
-    res.status(409).json({ error: msg });
-  } else if (msg.startsWith('Invalid') || msg.includes('required')) {
-    res.status(400).json({ error: msg });
-  } else {
-    res.status(500).json({ error: msg });
-  }
-}
-
-/**
- * Return the id of a live auto_review task pointing at this source — either
- * keyed on `pr_number` (poller-created) or on `review_of_task_id` (manual).
- * Used by both GET /api/tasks/:id and the manual-trigger endpoint.
- */
-function lookupExistingReviewId(task: { id: string; pr_number: number | null }): string | null {
-  if (task.pr_number != null) {
-    const byPr = findReviewTaskByPrNumber(task.pr_number);
-    if (byPr) return byPr.id;
-  }
-  const byLink = findReviewTaskBySource(task.id);
-  return byLink?.id ?? null;
-}
-
-/** Recursively merge `incoming` into `base` (objects merged, primitives overwritten). */
-function deepMerge(
-  base: Record<string, unknown>,
-  incoming: Record<string, unknown>,
-): Record<string, unknown> {
-  const result: Record<string, unknown> = { ...base };
-  for (const [key, val] of Object.entries(incoming)) {
-    if (
-      val !== null &&
-      typeof val === 'object' &&
-      !Array.isArray(val) &&
-      typeof result[key] === 'object' &&
-      result[key] !== null &&
-      !Array.isArray(result[key])
-    ) {
-      result[key] = deepMerge(
-        result[key] as Record<string, unknown>,
-        val as Record<string, unknown>,
-      );
-    } else {
-      result[key] = val;
-    }
-  }
-  return result;
-}
-
-/** Reload a task with its related agents and user_terminals. */
-function fetchTaskBundle(taskId: string): Task {
-  const task = getTaskRepo(taskId) as Task;
-  task.agents = listAllAgents(taskId);
-  task.user_terminals = listUserTerminals(taskId);
-  return task;
-}
-
-function derivedStatus(task: {
-  runtime_state: string;
-  agents: Array<{ status: string; hook_activity: string }>;
-}): DerivedTaskStatus | null {
-  if (task.runtime_state !== 'running') return null;
-  const activities = task.agents.filter((a) => a.status !== 'stopped').map((a) => a.hook_activity);
-  if (activities.length === 0) return 'done';
-  if (activities.includes('active')) return 'working';
-  if (activities.includes('waiting')) return 'needs_attention';
-  return 'done';
-}
-
-/** Flat Claude launch aliases for `/api/settings` responses (dashboard reads these keys). */
-function augmentDashboardSettings(settings: OctomuxSettings): OctomuxSettings & {
-  dangerouslySkipPermissions: boolean;
-  claudeFlags: string;
-} {
-  const cc = settings.harnesses?.['claude-code'] ?? {};
-  const dsp = (cc as { dangerouslySkipPermissions?: unknown }).dangerouslySkipPermissions;
-  const flagsRaw = (cc as { flags?: unknown }).flags;
-  return {
-    ...settings,
-    dangerouslySkipPermissions: typeof dsp === 'boolean' ? dsp : Boolean(dsp),
-    claudeFlags: typeof flagsRaw === 'string' ? flagsRaw : '',
-  };
-}
 
 export function setupRoutes(app: Express): void {
   app.use('/api/hooks', hookRoutes);
   mountArtifactEndpoint(app);
 
-  // GET /api/health — readiness probe: DB reachability, uptime, running tasks
-  app.get('/api/health', (_req: Request, res: Response) => {
-    const uptime = process.uptime();
-    const data_dir = getDataDir();
-
-    let db: { ok: true } | { ok: false; error: string };
-    let running_tasks = 0;
-    try {
-      pingDb();
-      db = { ok: true };
-      running_tasks = countRunningTasks();
-    } catch (err) {
-      db = { ok: false, error: err instanceof Error ? err.message : String(err) };
-    }
-
-    const status = db.ok ? 'ok' : 'degraded';
-    if (db.ok) {
-      healthLogger.info(
-        { operation: 'health', status, db_ok: true, running_tasks },
-        'health check',
-      );
-    } else {
-      healthLogger.warn(
-        { operation: 'health', status, db_ok: false, error: db.error },
-        'health check degraded',
-      );
-    }
-
-    res.status(db.ok ? 200 : 503).json({ status, uptime, db, running_tasks, data_dir });
-  });
-
-  // GET /api/harnesses — list registered harness implementations
-  app.get('/api/harnesses', (_req: Request, res: Response) => {
-    res.json(
-      listHarnesses().map(({ id, displayName, sessionIdMode }) => ({
-        id,
-        displayName,
-        sessionIdMode,
-      })),
-    );
-  });
-
-  // Browse directories for folder picker
-  app.get('/api/browse', async (req: Request, res: Response) => {
-    const dirPath = (req.query.path as string) || os.homedir();
-
-    try {
-      const stat = await fs.promises.stat(dirPath);
-      if (!stat.isDirectory()) {
-        res.status(400).json({ error: 'Path is not a directory' });
-        return;
-      }
-    } catch {
-      res.status(400).json({ error: 'Path does not exist' });
-      return;
-    }
-
-    const dirEntries = await fs.promises.readdir(dirPath);
-    const resolved = await Promise.all(
-      dirEntries.map(
-        async (name): Promise<{ name: string; path: string; isGit: boolean } | null> => {
-          const fullPath = path.join(dirPath, name);
-          try {
-            const stat = await fs.promises.stat(fullPath);
-            if (!stat.isDirectory()) return null;
-            const isGit = await fs.promises
-              .access(path.join(fullPath, '.git'))
-              .then(() => true)
-              .catch(() => false);
-            return { name, path: fullPath, isGit };
-          } catch {
-            return null;
-          }
-        },
-      ),
-    );
-    const entries = resolved.filter(
-      (e): e is { name: string; path: string; isGit: boolean } => e !== null,
-    );
-
-    entries.sort((a, b) => {
-      if (a.isGit !== b.isGit) return a.isGit ? -1 : 1;
-      const aHidden = a.name.startsWith('.');
-      const bHidden = b.name.startsWith('.');
-      if (aHidden !== bHidden) return aHidden ? 1 : -1;
-      return a.name.localeCompare(b.name);
-    });
-
-    const parent = path.dirname(dirPath);
-    res.json({
-      current: dirPath,
-      parent: parent !== dirPath ? parent : null,
-      entries,
-    });
-  });
-
-  // List branches for a git repo
-  app.get('/api/branches', async (req: Request, res: Response) => {
-    const repoPath = req.query.repo_path as string;
-    if (!repoPath) {
-      res.status(400).json({ error: 'repo_path is required' });
-      return;
-    }
-
-    try {
-      const { stdout } = await execFile('git', [
-        '-C',
-        repoPath,
-        'branch',
-        '-a',
-        '--format=%(refname:short)',
-      ]);
-      const branches = stdout
-        .trim()
-        .split('\n')
-        .filter(Boolean)
-        .map((b) => b.replace(/^origin\//, ''));
-      // Deduplicate (local + remote may overlap)
-      const unique = [...new Set(branches)].filter((b) => b !== 'HEAD');
-      res.json(unique);
-    } catch {
-      res.status(400).json({ error: 'Failed to list branches' });
-    }
-  });
-
-  // Preflight check for none-mode task creation
-  app.get('/api/preflight/none-mode', async (req: Request, res: Response) => {
-    const repoPath = String(req.query.repo_path ?? '');
-    const baseBranch = String(req.query.base_branch ?? '');
-    if (!repoPath || !baseBranch) {
-      res.status(400).json({ error: 'repo_path and base_branch are required' });
-      return;
-    }
-    try {
-      const { preflightNoneMode } = await import('./preflight.js');
-      const result = await preflightNoneMode(repoPath, baseBranch);
-      res.json(result);
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-    }
-  });
-
-  // Stash uncommitted changes before switching branch
-  app.post('/api/preflight/stash', async (req: Request, res: Response) => {
-    const repoPath = String(req.body?.repo_path ?? '');
-    const targetBranch = String(req.body?.target_branch ?? '');
-    if (!repoPath || !targetBranch) {
-      res.status(400).json({ error: 'repo_path and target_branch are required' });
-      return;
-    }
-    try {
-      await execFile('git', [
-        '-C',
-        repoPath,
-        'stash',
-        'push',
-        '-u',
-        '-m',
-        `octomux: auto-stash before switching to ${targetBranch}`,
-      ]);
-      res.json({ ok: true });
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-    }
-  });
-
-  // Get default branch for a git repo
-  app.get('/api/default-branch', async (req: Request, res: Response) => {
-    const repoPath = req.query.repo_path as string;
-    if (!repoPath) {
-      res.status(400).json({ error: 'repo_path is required' });
-      return;
-    }
-
-    try {
-      const { stdout } = await execFile('git', [
-        '-C',
-        repoPath,
-        'symbolic-ref',
-        'refs/remotes/origin/HEAD',
-      ]);
-      const branch = stdout.trim().replace('refs/remotes/origin/', '');
-      res.json({ branch });
-    } catch {
-      // Fallback to 'main'
-      res.json({ branch: 'main' });
-    }
-  });
-
-  // Recent repository paths from past tasks
-  app.get('/api/recent-repos', (_req: Request, res: Response) => {
-    res.json(listRecentRepoPaths(10));
-  });
+  // Mount extracted routers (bare app.use — each router keeps full /api/... paths)
+  app.use(miscRouter);
+  app.use(learningsRouter);
+  app.use(skillsRouter);
+  app.use(teamsRouter);
+  app.use(setupRouter);
+  app.use(settingsRouter);
+  app.use(hooksRegistryRouter);
+  app.use(chatsRouter);
+  app.use(agentDefsRouter);
+  app.use(orchestratorRouter);
+  app.use(integrationsRouter);
 
   // List all tasks with agents
   app.get('/api/tasks', (req: Request, res: Response) => {
@@ -1650,378 +1326,6 @@ export function setupRoutes(app: Express): void {
     res.status(204).send();
   });
 
-  // ─── Setup / onboarding ─────────────────────────────────────────────────────
-
-  app.get('/api/setup/status', async (_req: Request, res: Response) => {
-    try {
-      const { getSetupStatus } = await import('./setup-status.js');
-      res.json(await getSetupStatus());
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.post('/api/setup/install', async (req: Request, res: Response) => {
-    const { id } = req.body as { id?: string };
-    if (!id || typeof id !== 'string') {
-      res.status(400).json({ error: 'body must contain { id: string }' });
-      return;
-    }
-    try {
-      const { runSetupInstall } = await import('./setup-status.js');
-      const result = await runSetupInstall(id);
-      res.json(result);
-    } catch (err) {
-      const message = (err as Error).message;
-      if (message.startsWith('Install not allowed')) {
-        res.status(400).json({ error: message });
-      } else {
-        res.status(500).json({ error: message });
-      }
-    }
-  });
-
-  app.post('/api/setup/apply-recommended-defaults', async (_req: Request, res: Response) => {
-    try {
-      const { applyRecommendedDefaults } = await import('./setup-status.js');
-      const settings = await applyRecommendedDefaults();
-      res.json(augmentDashboardSettings(settings));
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/hooks/templates', async (_req: Request, res: Response) => {
-    try {
-      const { listHookTemplates, isHookTemplateInstalled } = await import('./hooks-install.js');
-      const templates = listHookTemplates().map((id) => ({
-        id,
-        installed: isHookTemplateInstalled(id),
-      }));
-      res.json(templates);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.post('/api/hooks/install', async (req: Request, res: Response) => {
-    const { template } = req.body as { template?: string };
-    if (!template || typeof template !== 'string') {
-      res.status(400).json({ error: 'body must contain { template: string }' });
-      return;
-    }
-    try {
-      const { installHookTemplate } = await import('./hooks-install.js');
-      const files = installHookTemplate(template);
-      res.json({ ok: true, files });
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/settings', async (_req: Request, res: Response) => {
-    try {
-      const settings = await getSettings();
-      const envClaudeFlags = process.env.OCTOMUX_CLAUDE_FLAGS?.trim();
-      res.json({
-        ...augmentDashboardSettings(settings),
-        envOverrides: {
-          claudeFlags: envClaudeFlags ? envClaudeFlags : null,
-        },
-      });
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.patch('/api/settings', async (req: Request, res: Response) => {
-    try {
-      const settings = await updateSettings(req.body);
-      res.json(augmentDashboardSettings(settings));
-    } catch (err) {
-      const message = (err as Error).message;
-      const clientInputError =
-        message.startsWith('Invalid editor') ||
-        message.startsWith('Invalid claudeFlags') ||
-        message.includes('Invalid claude-code') ||
-        message.includes('Invalid harnesses.claude-code');
-      if (clientInputError) {
-        res.status(400).json({ error: message });
-      } else {
-        res.status(500).json({ error: message });
-      }
-    }
-  });
-
-  // ─── Repo Config ────────────────────────────────────────────────────────────
-
-  app.get('/api/repo-configs', async (_req: Request, res: Response) => {
-    try {
-      const configs = listRepoConfigs();
-      res.json(configs);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/repo-config', async (req: Request, res: Response) => {
-    const repoPath = req.query.repo_path as string;
-    if (!repoPath) {
-      res.status(400).json({ error: 'repo_path query parameter is required' });
-      return;
-    }
-    try {
-      const config = await getOrCreateRepoConfig(repoPath);
-      res.json(config);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.patch('/api/repo-config', async (req: Request, res: Response) => {
-    const { repo_path, ...updates } = req.body as Record<string, unknown>;
-    if (!repo_path || typeof repo_path !== 'string') {
-      res.status(400).json({ error: 'repo_path is required' });
-      return;
-    }
-    try {
-      const config = updateRepoConfig(repo_path, updates);
-      res.json(config);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // ─── Agents ──────────────────────────────────────────────────────────────────
-
-  app.get('/api/agents', async (_req: Request, res: Response) => {
-    try {
-      const agents = await listAgents();
-      res.json(agents);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/agents/:name', async (req: Request, res: Response) => {
-    try {
-      const agent = await getAgent(req.params.name as string);
-      res.json(agent);
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  app.put('/api/agents/:name', async (req: Request, res: Response) => {
-    const { content } = req.body as { content?: string };
-    if (content === undefined || content === null) {
-      res.status(400).json({ error: 'content is required' });
-      return;
-    }
-    try {
-      await saveAgent(req.params.name as string, content);
-      await syncAgents();
-      const agent = await getAgent(req.params.name as string);
-      res.json(agent);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.delete('/api/agents/:name', async (req: Request, res: Response) => {
-    try {
-      const name = req.params.name as string;
-      if (isBuiltInAgent(name)) {
-        await resetAgent(name);
-      } else {
-        await deleteAgent(name);
-      }
-      await syncAgents();
-      res.json({ ok: true });
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  app.post('/api/agents', async (req: Request, res: Response) => {
-    const { name, content } = req.body as { name?: string; content?: string };
-    if (!name || !content) {
-      res.status(400).json({ error: 'name and content are required' });
-      return;
-    }
-    try {
-      await createAgent(name, content);
-      await syncAgents();
-      const agent = await getAgent(name);
-      res.status(201).json(agent);
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  // ─── Chats (standalone runtime agents) ───────────────────────────────────
-  // Distinct from /api/agents (agent-prompt definitions). A "chat" is a
-  // tmux-backed Claude runtime instance with `agents.task_id = NULL`.
-
-  app.get('/api/chats', (_req: Request, res: Response) => {
-    try {
-      res.json(listChats());
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/chats/:id', (req: Request, res: Response) => {
-    const chat = getChat(req.params.id as string);
-    if (!chat) {
-      res.status(404).json({ error: 'Chat not found' });
-      return;
-    }
-    res.json(chat);
-  });
-
-  /**
-   * Move a runtime agent between task ids (or detach to a standalone chat
-   * with task_id=null). Kills the old tmux window, opens a new one at the new
-   * cwd, and resumes claude so transcript context survives.
-   */
-  app.patch('/api/agents/:id/task', async (req: Request, res: Response) => {
-    const agent = getAgentRepo(req.params.id as string);
-    if (!agent) {
-      res.status(404).json({ error: 'Agent not found' });
-      return;
-    }
-    const body = (req.body ?? {}) as { task_id?: string | null };
-    if (!('task_id' in body)) {
-      res.status(400).json({ error: 'task_id is required (string or null)' });
-      return;
-    }
-    const targetTaskId = body.task_id;
-    if (targetTaskId !== null && typeof targetTaskId !== 'string') {
-      res.status(400).json({ error: 'task_id must be a string or null' });
-      return;
-    }
-    if (targetTaskId === agent.task_id) {
-      res.status(400).json({ error: 'Agent is already on that task' });
-      return;
-    }
-    if (targetTaskId !== null) {
-      const targetTask = getTaskRepo(targetTaskId);
-      if (!targetTask) {
-        res.status(404).json({ error: `Task not found: ${targetTaskId}` });
-        return;
-      }
-      const trs = (targetTask as Task).runtime_state;
-      if (!(['setting_up', 'running'] as const).includes(trs as 'running')) {
-        res.status(409).json({ error: `Target task is not active (runtime_state=${trs})` });
-        return;
-      }
-      if (!targetTask.worktree_id) {
-        res.status(409).json({ error: 'Target task has no worktree' });
-        return;
-      }
-    }
-
-    try {
-      const updated = await hopAgent(agent, targetTaskId);
-      broadcast({ type: 'task:updated', payload: { taskId: agent.task_id ?? targetTaskId ?? '' } });
-      res.json(updated);
-    } catch (err) {
-      const msg = (err as Error).message;
-      apiLogger.error(
-        {
-          agent_id: agent.id,
-          from_task_id: agent.task_id,
-          to_task_id: targetTaskId,
-          operation: 'task_hop',
-          err,
-        },
-        'task_hop: failed',
-      );
-      if (msg.includes('not found') || msg.includes('does not exist')) {
-        res.status(409).json({ error: msg });
-      } else {
-        res.status(500).json({ error: msg });
-      }
-    }
-  });
-
-  app.post('/api/chats', async (req: Request, res: Response) => {
-    const body = (req.body ?? {}) as CreateChatRequest;
-
-    if (body.agent != null) {
-      try {
-        validateAgentName(body.agent);
-      } catch (err) {
-        res.status(400).json({ error: (err as Error).message });
-        return;
-      }
-    }
-
-    if (body.harness_id != null) {
-      try {
-        getHarness(body.harness_id);
-      } catch (err) {
-        res.status(400).json({ error: (err as Error).message });
-        return;
-      }
-    }
-
-    try {
-      const chat = await createChat({
-        label: body.label,
-        cwd: body.cwd,
-        agent: body.agent,
-        prompt: body.prompt,
-        harnessId: body.harness_id,
-      });
-      res.status(201).json(chat);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  /**
-   * Close a chat — stop the tmux session, mark the agent stopped.
-   * Body: `{ status: 'stopped' }`. Preserves the row so history stays visible.
-   */
-  app.patch('/api/chats/:id', async (req: Request, res: Response) => {
-    const chat = getChat(req.params.id as string);
-    if (!chat) {
-      res.status(404).json({ error: 'Chat not found' });
-      return;
-    }
-    const body = (req.body ?? {}) as { status?: string };
-    if (body.status !== 'stopped') {
-      res.status(400).json({ error: "Only status='stopped' is supported" });
-      return;
-    }
-    try {
-      await closeChat(chat);
-      const updated = getChat(chat.id);
-      broadcast({ type: 'chat:updated', payload: { chatId: chat.id } });
-      res.json(updated);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  /** Delete a chat — kill tmux, remove scratch dir, delete DB row. */
-  app.delete('/api/chats/:id', async (req: Request, res: Response) => {
-    const chat = getChat(req.params.id as string);
-    if (!chat) {
-      res.status(404).json({ error: 'Chat not found' });
-      return;
-    }
-    try {
-      await deleteChat(chat);
-      broadcast({ type: 'chat:deleted', payload: { chatId: chat.id } });
-      res.status(204).send();
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
   // ─── Worktrees (browser) ─────────────────────────────────────────────────
 
   app.get('/api/worktrees', (_req: Request, res: Response) => {
@@ -2363,454 +1667,6 @@ export function setupRoutes(app: Express): void {
     res.json(executions);
   });
 
-  // ─── Teams ───────────────────────────────────────────────────────────────────
-
-  app.post('/api/teams/run', async (req: Request, res: Response) => {
-    const { name, repo_path } = req.body as { name?: string; repo_path?: string };
-    if (!name) {
-      res.status(400).json({ error: 'name is required' });
-      return;
-    }
-    if (!repo_path) {
-      res.status(400).json({ error: 'repo_path is required' });
-      return;
-    }
-    try {
-      const { runTeam } = await import('./teams.js');
-      const taskId = await runTeam({ name, repoPath: repo_path });
-      res.status(201).json({ task_id: taskId });
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-    }
-  });
-
-  app.post('/api/teams/schedule', async (req: Request, res: Response) => {
-    const { name, repo_path, cron } = req.body as {
-      name?: string;
-      repo_path?: string;
-      cron?: string;
-    };
-    if (!name) {
-      res.status(400).json({ error: 'name is required' });
-      return;
-    }
-    if (!repo_path) {
-      res.status(400).json({ error: 'repo_path is required' });
-      return;
-    }
-    if (!cron) {
-      res.status(400).json({ error: 'cron is required' });
-      return;
-    }
-    try {
-      const { upsertTeamSchedule } = await import('./teams.js');
-      upsertTeamSchedule({ name, repoPath: repo_path, cron });
-      res.status(200).json({ ok: true });
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/teams', async (_req: Request, res: Response) => {
-    try {
-      const { listTeamSchedules } = await import('./teams.js');
-      const schedules = listTeamSchedules();
-      res.json(schedules);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // ─── Skills ──────────────────────────────────────────────────────────────────
-
-  app.get('/api/skills', async (_req: Request, res: Response) => {
-    try {
-      const skills = await listSkills();
-      res.json(skills);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/skills/:name', async (req: Request, res: Response) => {
-    try {
-      const skill = await getSkill(req.params.name as string);
-      res.json(skill);
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  app.post('/api/skills', async (req: Request, res: Response) => {
-    const { name, content } = req.body as { name?: string; content?: string };
-    if (!name) {
-      res.status(400).json({ error: 'name is required' });
-      return;
-    }
-    try {
-      const skill = await createSkill(name, content || '');
-      res.status(201).json(skill);
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  app.put('/api/skills/:name', async (req: Request, res: Response) => {
-    const { content } = req.body as { content?: string };
-    if (content === undefined || content === null) {
-      res.status(400).json({ error: 'content is required' });
-      return;
-    }
-    try {
-      const skill = await updateSkill(req.params.name as string, content);
-      res.json(skill);
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  app.delete('/api/skills/:name', async (req: Request, res: Response) => {
-    try {
-      await deleteSkill(req.params.name as string);
-      res.status(204).send();
-    } catch (err) {
-      sendDomainError(res, err);
-    }
-  });
-
-  // ─── Integrations ─────────────────────────────────────────────────────────────
-
-  // GET /api/integrations/providers — list registered providers with their schemas
-  app.get('/api/integrations/providers', (_req: Request, res: Response) => {
-    const providers = listProviders().map((p) => ({
-      kind: p.kind,
-      displayName: p.displayName,
-      configSchema: p.configSchema,
-      events: p.events,
-    }));
-    res.json(providers);
-  });
-
-  // GET /api/integrations — list all configured integrations (config masked)
-  app.get('/api/integrations', (_req: Request, res: Response) => {
-    const integrations = listIntegrations();
-    const masked = integrations.map((i) => {
-      const provider = getProvider(i.kind);
-      if (!provider) return i;
-      return maskIntegration(i, provider.configSchema);
-    });
-    res.json(masked);
-  });
-
-  // POST /api/integrations — create a new integration
-  app.post('/api/integrations', (req: Request, res: Response) => {
-    const body = req.body as { kind?: string; name?: string; config?: unknown };
-    if (!body.kind?.trim()) {
-      res.status(400).json({ error: 'kind is required' });
-      return;
-    }
-    if (!body.name?.trim()) {
-      res.status(400).json({ error: 'name is required' });
-      return;
-    }
-    const provider = getProvider(body.kind);
-    if (!provider) {
-      res.status(400).json({ error: `unknown integration kind: ${body.kind}` });
-      return;
-    }
-    const validation = provider.validate(body.config ?? {});
-    if (!validation.ok) {
-      res.status(400).json({ error: 'config validation failed', details: validation.errors });
-      return;
-    }
-    const integration = createIntegration(body.kind, body.name, body.config ?? {});
-    res.status(201).json(maskIntegration(integration, provider.configSchema));
-  });
-
-  // PATCH /api/integrations/:id — update an integration
-  app.patch('/api/integrations/:id', (req: Request, res: Response) => {
-    const id = (req.params as Record<string, string>).id;
-    const existing = getIntegration(id);
-    if (!existing) {
-      res.status(404).json({ error: 'Integration not found' });
-      return;
-    }
-    const provider = getProvider(existing.kind);
-    const body = req.body as { name?: string; config?: unknown; enabled?: boolean };
-    const patch: { name?: string; config?: unknown; enabled?: boolean } = {};
-
-    if (body.name !== undefined) patch.name = body.name;
-    if (body.enabled !== undefined) patch.enabled = body.enabled;
-
-    if (body.config !== undefined) {
-      const mergedConfig = provider
-        ? mergeMaskedConfig(existing.config, body.config, provider.configSchema)
-        : (body.config as Record<string, unknown>);
-      const validation = provider ? provider.validate(mergedConfig) : { ok: true };
-      if (!validation.ok) {
-        res.status(400).json({
-          error: 'config validation failed',
-          details: (validation as { errors?: string[] }).errors,
-        });
-        return;
-      }
-      patch.config = mergedConfig;
-    }
-
-    const updated = updateIntegration(id, patch);
-    if (!updated) {
-      res.status(404).json({ error: 'Integration not found' });
-      return;
-    }
-    res.json(provider ? maskIntegration(updated, provider.configSchema) : updated);
-  });
-
-  // DELETE /api/integrations/:id
-  app.delete('/api/integrations/:id', (req: Request, res: Response) => {
-    const id = (req.params as Record<string, string>).id;
-    const existing = getIntegration(id);
-    if (!existing) {
-      res.status(404).json({ error: 'Integration not found' });
-      return;
-    }
-    deleteIntegration(id);
-    res.status(204).send();
-  });
-
-  // POST /api/integrations/linear/prefill — fetch teams/states and build a prefilled config map
-  app.post('/api/integrations/linear/prefill', async (req: Request, res: Response) => {
-    const body = req.body as { api_key?: string };
-    const apiKey = body.api_key?.trim();
-    if (!apiKey) {
-      res.status(400).json({ error: 'api_key is required' });
-      return;
-    }
-    try {
-      const { prefillFromLinear } = await import('./integrations/linear/prefill.js');
-      const result = await prefillFromLinear(apiKey);
-      res.json(result);
-    } catch (err) {
-      const message = (err as Error).message;
-      res.status(502).json({ error: message });
-    }
-  });
-
-  // POST /api/integrations/:id/test — test the connection using stored (unmasked) config
-  app.post('/api/integrations/:id/test', async (req: Request, res: Response) => {
-    const id = (req.params as Record<string, string>).id;
-    const existing = getIntegration(id);
-    if (!existing) {
-      res.status(404).json({ error: 'Integration not found' });
-      return;
-    }
-    const provider = getProvider(existing.kind);
-    if (!provider) {
-      res.status(400).json({ error: `no provider for kind: ${existing.kind}` });
-      return;
-    }
-    if (!provider.test) {
-      res.json({ ok: true, message: 'Provider does not support connection testing' });
-      return;
-    }
-    try {
-      const result = await provider.test(existing.config);
-      res.json(result);
-    } catch (err) {
-      res.status(500).json({ ok: false, message: (err as Error).message });
-    }
-  });
-
-  // ─── C4: Hook registry endpoints ──────────────────────────────────────────
-
-  const ALL_HOOK_EVENTS = [
-    'workflow_status_changed',
-    'summary_updated',
-    'note_added',
-    'ref_added',
-    'ref_removed',
-    'task_created',
-    'runtime_state_changed',
-  ] as const;
-
-  const isProduction = process.env.NODE_ENV === 'production';
-  const hooksLogsDir = isProduction
-    ? path.join(octomuxRoot(), 'logs', 'hooks')
-    : path.join(__dirname, '..', 'data', 'logs', 'hooks');
-
-  interface HookRegistryEntry {
-    scope: 'global' | `repo:${string}` | 'builtin';
-    key: string;
-    event: string | null;
-    script_path: string | null;
-    description: string | null;
-    enabled: boolean;
-    requires_env: string | null;
-    last_run_at: string | null;
-    last_exit_code: number | null;
-  }
-
-  /** Parse the most-recent log file for a given event+script-basename. */
-  function findLastRunMeta(
-    event: string,
-    scriptName: string,
-  ): { last_run_at: string | null; last_exit_code: number | null } {
-    try {
-      if (!fs.existsSync(hooksLogsDir)) return { last_run_at: null, last_exit_code: null };
-      const prefix = `${event}-`;
-      const suffix = `-${scriptName}`;
-      const files = fs
-        .readdirSync(hooksLogsDir)
-        .filter(
-          (f) => f.startsWith(prefix) && (f.endsWith(`${suffix}.log`) || f.includes(`${suffix}-`)),
-        )
-        .map((f) => {
-          try {
-            return { f, mtime: fs.statSync(path.join(hooksLogsDir, f)).mtimeMs };
-          } catch {
-            return { f, mtime: 0 };
-          }
-        })
-        .sort((a, b) => b.mtime - a.mtime);
-
-      if (files.length === 0) return { last_run_at: null, last_exit_code: null };
-
-      const logPath = path.join(hooksLogsDir, files[0].f);
-      const content = fs.readFileSync(logPath, 'utf8');
-      const lines = content.split('\n');
-      const headerLine = lines.find((l) => l.startsWith('[octomux] event='));
-      const footerLine = lines.slice(1).find((l) => l.startsWith('[octomux] duration_ms='));
-
-      let last_run_at: string | null = null;
-      let last_exit_code: number | null = null;
-
-      if (headerLine) {
-        const m = headerLine.match(/started_at=(\d+)/);
-        if (m) last_run_at = new Date(parseInt(m[1], 10)).toISOString();
-      }
-      if (footerLine) {
-        const ec = footerLine.match(/exit_code=(-?\d+)/);
-        if (ec) last_exit_code = parseInt(ec[1], 10);
-      }
-      return { last_run_at, last_exit_code };
-    } catch {
-      return { last_run_at: null, last_exit_code: null };
-    }
-  }
-
-  /** Read enabled state from hook_settings; missing row = defaultEnabled. */
-  function getHookEnabled(scope: string, key: string, defaultEnabled: boolean): boolean {
-    return getHookEnabledRepo(scope, key, defaultEnabled);
-  }
-
-  /** Discover scripts for all events under a hooks base directory. */
-  function discoverHookScripts(
-    hooksBase: string,
-    scope: HookRegistryEntry['scope'],
-  ): HookRegistryEntry[] {
-    const entries: HookRegistryEntry[] = [];
-    for (const event of ALL_HOOK_EVENTS) {
-      const dir = path.join(hooksBase, `${event}.d`);
-      try {
-        if (!fs.existsSync(dir)) continue;
-        const files = fs
-          .readdirSync(dir)
-          .filter((name) => {
-            try {
-              return fs.statSync(path.join(dir, name)).isFile();
-            } catch {
-              return false;
-            }
-          })
-          .sort();
-        for (const name of files) {
-          const key = `${event}/${name}`;
-          const runMeta = findLastRunMeta(event, name);
-          entries.push({
-            scope,
-            key,
-            event,
-            script_path: path.join(dir, name),
-            description: null,
-            enabled: getHookEnabled(scope, key, true),
-            requires_env: null,
-            last_run_at: runMeta.last_run_at,
-            last_exit_code: runMeta.last_exit_code,
-          });
-        }
-      } catch {
-        // skip unreadable dirs
-      }
-    }
-    return entries;
-  }
-
-  // GET /api/hooks/registry — list all hooks with enabled state
-  app.get('/api/hooks/registry', (_req: Request, res: Response) => {
-    const entries: HookRegistryEntry[] = [];
-
-    // Built-in: summarize-progress (defaults disabled)
-    const builtinEnabled = getHookEnabled('builtin', 'summarize-progress', false);
-    entries.push({
-      scope: 'builtin',
-      key: 'summarize-progress',
-      event: null,
-      script_path: null,
-      description:
-        'After each agent stop, calls Haiku to write a one-sentence progress summary to tasks.current_summary.',
-      enabled: builtinEnabled,
-      requires_env: process.env.ANTHROPIC_API_KEY ? null : 'ANTHROPIC_API_KEY',
-      last_run_at: null,
-      last_exit_code: null,
-    });
-
-    // Global hooks: ~/.octomux/hooks/
-    const globalHooksBase = path.join(octomuxRoot(), 'hooks');
-    entries.push(...discoverHookScripts(globalHooksBase, 'global'));
-
-    // Repo hooks: collect from every active task's repo_path
-    try {
-      const activeTasks = listActiveRepoPaths();
-
-      const seen = new Set<string>();
-      for (const { repo_path } of activeTasks) {
-        if (seen.has(repo_path)) continue;
-        seen.add(repo_path);
-        const scope: HookRegistryEntry['scope'] = `repo:${repo_path}`;
-        const repoHooksBase = path.join(repo_path, '.octomux', 'hooks');
-        entries.push(...discoverHookScripts(repoHooksBase, scope));
-      }
-    } catch {
-      // DB error — skip repo hooks
-    }
-
-    res.json({ hooks: entries });
-  });
-
-  // PATCH /api/hooks/registry/:scope/:key — toggle a hook
-  app.patch('/api/hooks/registry/:scope/:key', (req: Request, res: Response) => {
-    const params = req.params as Record<string, string>;
-    const scope = decodeURIComponent(params.scope);
-    const key = decodeURIComponent(params.key);
-    const { enabled } = req.body as { enabled?: unknown };
-
-    if (typeof enabled !== 'boolean') {
-      res.status(400).json({ error: 'body must contain { enabled: boolean }' });
-      return;
-    }
-
-    try {
-      upsertHookSetting(scope, key, enabled);
-
-      // Invalidate dispatcher cache for this entry
-      invalidateHookEnabledCache(scope, key);
-
-      res.json({ scope, key, enabled });
-    } catch (err) {
-      apiLogger.warn({ scope, key, err }, 'failed to update hook_settings');
-      res.status(500).json({ error: 'failed to update hook setting' });
-    }
-  });
-
   // ─── Reviews inbox ───────────────────────────────────────────────────────────
 
   // GET /api/reviews — list all auto_review tasks with aggregated counts
@@ -3088,159 +1944,6 @@ export function setupRoutes(app: Express): void {
     });
 
     res.status(201).json({ id: newId, action: 'created' });
-  });
-
-  // ─── Orchestrator chat ───────────────────────────────────────────────────────
-
-  // POST /api/orchestrator/conversations — create a new orchestrator conversation
-  app.post('/api/orchestrator/conversations', async (req: Request, res: Response) => {
-    const { title, cwd } = req.body as { title?: string; cwd?: string };
-    if (!title?.trim()) {
-      res.status(400).json({ error: 'title is required' });
-      return;
-    }
-    try {
-      const id = createConversation({ title: title.trim() });
-      // The conductor runs in a trusted cwd (default: the server's repo root).
-      const convCwd = cwd?.trim() || process.cwd();
-      // Launch the interactive claude session for this conversation (tmux + transcript).
-      await startConversation(id, convCwd);
-      apiLogger.info(
-        { conversation_id: id, operation: 'createConversation', cwd: convCwd },
-        'orchestrator conversation created + session launched',
-      );
-      const conv = getConversation(id);
-      res.status(201).json(conv);
-    } catch (err) {
-      apiLogger.error(
-        { err, operation: 'createConversation' },
-        'failed to create orchestrator conversation',
-      );
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // GET /api/orchestrator/conversations — list all conversations
-  app.get('/api/orchestrator/conversations', (_req: Request, res: Response) => {
-    try {
-      res.json(listConversations());
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // GET /api/orchestrator/conversations/:id — get a single conversation
-  app.get('/api/orchestrator/conversations/:id', (req: Request, res: Response) => {
-    try {
-      const conv = getConversation((req.params as Record<string, string>).id);
-      if (!conv) {
-        res.status(404).json({ error: 'Conversation not found' });
-        return;
-      }
-      res.json(conv);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // GET /api/orchestrator/conversations/:id/messages — list messages for a conversation
-  app.get('/api/orchestrator/conversations/:id/messages', (req: Request, res: Response) => {
-    try {
-      const convId = (req.params as Record<string, string>).id;
-      const conv = getConversation(convId);
-      if (!conv) {
-        res.status(404).json({ error: 'Conversation not found' });
-        return;
-      }
-      res.json(listOrchestratorMessages(convId));
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // POST /api/orchestrator/conversations/:id/global-monitor — toggle global-monitor mode
-  // Exactly one conversation may be in global-monitor mode at a time.
-  // If the conversation is already the global monitor, clears it.
-  // Otherwise, designates it as the global monitor (clearing the previous one).
-  app.post('/api/orchestrator/conversations/:id/global-monitor', (req: Request, res: Response) => {
-    try {
-      const convId = (req.params as Record<string, string>).id;
-      const conv = getConversation(convId);
-      if (!conv) {
-        res.status(404).json({ error: 'Conversation not found' });
-        return;
-      }
-      // Toggle: if already global-monitor, clear; otherwise set
-      const currentMonitor = getGlobalMonitorConversation();
-      let isMonitor: boolean;
-      if (currentMonitor === convId) {
-        clearGlobalMonitor();
-        isMonitor = false;
-      } else {
-        setGlobalMonitor(convId);
-        isMonitor = true;
-      }
-      apiLogger.info(
-        { conversation_id: convId, is_global_monitor: isMonitor },
-        'orchestrator: global-monitor toggled',
-      );
-      res.json({ is_global_monitor: isMonitor });
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // GET /api/orchestrator/conversations/:id/usage — conductor-leanness stats (§6.7)
-  // Returns tasks_spawned, tool_calls, started_at, last_activity_at.
-  // Returns zeros when no usage row exists yet (conversation was created but no
-  // write-actions have been dispatched).
-  app.get('/api/orchestrator/conversations/:id/usage', (req: Request, res: Response) => {
-    try {
-      const convId = (req.params as Record<string, string>).id;
-      const conv = getConversation(convId);
-      if (!conv) {
-        res.status(404).json({ error: 'Conversation not found' });
-        return;
-      }
-      const usage = getConversationUsage(convId);
-      if (!usage) {
-        // No usage row yet — return zeros so the UI always gets a valid shape.
-        res.json({
-          conversation_id: convId,
-          tasks_spawned: 0,
-          tool_calls: 0,
-          started_at: conv.created_at,
-          last_activity_at: conv.updated_at,
-        });
-        return;
-      }
-      res.json(usage);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // ─── Review learnings ────────────────────────────────────────────────────────
-
-  // GET /api/repos/:repoPath/learnings — list learnings for a repo
-  app.get('/api/repos/:repoPath/learnings', (req: Request, res: Response) => {
-    const repoPath = decodeURIComponent((req.params as Record<string, string>).repoPath);
-    try {
-      res.json(listLearningsForRepo(repoPath));
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // DELETE /api/learnings/:id — delete a single learning
-  app.delete('/api/learnings/:id', (req: Request, res: Response) => {
-    const id = (req.params as Record<string, string>).id;
-    try {
-      deleteLearning(id);
-      res.status(204).send();
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
   });
 
   // ─── Test-only seed endpoint ─────────────────────────────────────────────────
